SQL-507: Update *DatabaseMetaData classes
